Separated out changeEvents functions by page
