+++ src/js/questions/changeEvents/index.js
@@ -1,0 +1,9 @@
+
+/* eslint no-console: "off" */
+import careerPlansPage from './careerPlansPage'
+import initialPlage from './initialPage'
+
+export default {
+	...initialPlage,
+	...careerPlansPage
+}
