+++ src/views/layouts/main.handlebars
@@ -5,10 +5,10 @@
 	<meta charset="utf-8">
 	<title>Life Events</title>
 	<meta name="viewport" content="width=device-width, initial-scale=1">
-<link href="../../../dist/app.155e43cbfbe9ca7b232e.css" rel="stylesheet"></head>
+<link href="../../../dist/app.963620744585546996a2.css" rel="stylesheet"></head>
 
 <body style="margin: 0; background-color: #eeeeee;">
 	{{{body}}}
-<script type="text/javascript" src="../../../dist/app.155e43cbfbe9ca7b232e.js"></script></body>
+<script type="text/javascript" src="../../../dist/app.963620744585546996a2.js"></script></body>
 
 </html>
